demos/eml_training/requirements.txt
functions-framework==3.*
# venumML==0.2.*
google-cloud-pubsub==2.*
google-cloud-storage==2.*
requests==2.*